Move visualization to helper and downsample data array
